feat: Implement the metrics of usage for each provider (#38)
* feat: Implement the metrics of usage for each provider

Still need to decide if I decouple usage from latency. Because those are two separate things.


* refactor: Simplify the if-else with switch statement


* fix(test): Add usage metrics in ollama only to non-streaming responses


* fix: Log error when failing to decode request in GenerateProvidersTokenHandler


* fix(test): Log error when failing to decode request in TestGenerateProvidersTokenHandler


* refactor: Add debug logging for stream chunk processing in handleStreamingRequest


* refactor: Remove unnecessary else block for telemetry middleware initialization

I only use the middleware if it's enabled, i.e no need an noop instance.


* docs(examples): Add Kubernetes monitoring setup with Inference Gateway, OpenTelemetry, Prometheus, and Grafana

Partially implemented, will continue with it probably tomorrow.


* build: Add Helm installation to dev container


* chore: Update curl commands to use -fsSL for consistency


* docs(monitoring): Add Kubernetes cluster setup with Grafana Operator

Use a more idempotent approach with ctlptl and make it more declarative  for clarity.


* docs(monitoring): Update Taskfile to use helm upgrade for Grafana Operator installation


* feat(telemetry): Add Prometheus metrics server and record latency in telemetry middleware

Use Pull-based approach, the application will push metrics to local prometheus server and otel collector will pull the metrics in order to export to further systems, this way there is a lot of flexibility regarding to the metrics usage.


* docs(monitoring): Use a more declarative approach for clarity


* chore: Update examples/kubernetes/monitoring/README.md

* fix(telemetry): Update telemetry initialization and metrics server handling


* feat(monitoring): Enhance Inference Gateway deployment with Prometheus metrics and ServiceMonitor

Those are the only components that are needed, keeping it is simple for clarity reasons. Collector would only make sense when using push-based approach or when exporting to different instances is needed. WIll keep the example super simple.


* docs(grafana): Remove ingress, just proxy the service locally for the example


* docs(monitoring): Update Prometheus configuration and RBAC for Inference Gateway integration


* fix(grafana): Update Prometheus datasource URL to use the operated instance


* fix(telemetry): Correct parameter order for latency and token usage recording


* fix(grafana): Update Prometheus queries to use regex for metric names


* docs(readme): Correct typo in grafana dashboard access instructions


* docs(readme): Update Inference Gateway proxy instructions in monitoring README


* docs(readme): Add link to Inference Gateway Metrics dashboard in monitoring README


* fix(otel): Use application name for meter initialization instead of hardcoded value


* docs(readme): Update monitoring setup instructions and remove Taskfile

Less abstraction is clearer to understand and maintain.


* docs(readme): Update monitoring example to specify Prometheus Exporter and adjust prerequisites formatting


* docs(monitoring): Update Inference Gateway Metrics link to use markdown formatting


* feat(devcontainer): Add ctlptl completion to zsh configuration


* docs(monitoring): Update README to include testing multiple provider models with curl


* fix(monitoring): Update curl command to use silent mode for cleaner output


* refactor(monitoring): Update metric names and expressions for consistency in Grafana dashboards and OpenTelemetry integration


* docs(examples): Improve monitoring panel to show percentage of the distributed prompt tokens across providers


* refactor(monitoring): Fix the total tokens by providers display


* refactor(monitoring): Enhance Grafana dashboard for prompt token usage and distribution


* feat(telemetry): Add metrics for request count, response status, and request duration

Also add a panel to display error rate per provider.


* chore: Cleanup


* refactor(monitoring): Update Grafana dashboard expressions to filter by provider


* feat(monitoring): Add Grafana panel for Tokens per Second by Provider


* feat(monitoring): Add Request Latency panel to Grafana dashboard by Provider


* feat(monitoring): Add Ollama deployment and service configuration for local LLM models

Just to see how slow it is in comparison to cloud providers.

Running locally is obviously slower, one need strong hardware to run it locally. I use only tinyllama for demonstration purposes.


* docs: FIx typo


---------

+++ providers/common_types.go
@@ -138,11 +138,11 @@ type ResponseTokens struct {
 
 type Usage struct {
 	QueueTime        float64 `json:"queue_time"`
-	PromptTokens     int     `json:"prompt_tokens"`
+	PromptTokens     int64   `json:"prompt_tokens"`
 	PromptTime       float64 `json:"prompt_time"`
-	CompletionTokens int     `json:"completion_tokens"`
+	CompletionTokens int64   `json:"completion_tokens"`
 	CompletionTime   float64 `json:"completion_time"`
-	TotalTokens      int     `json:"total_tokens"`
+	TotalTokens      int64   `json:"total_tokens"`
 	TotalTime        float64 `json:"total_time"`
 }
